Created fastapi script that should work with
So generated a load of stuff because I reran the entire training of the model and generated a csv for use elsewhere and an image. the model seems to work properly. then I started to put together the working FastAPI implementation file. The functions work as tested in the notebook but will need to further test the file with the actual endpoint